js css fixes
