+++ index.html
@@ -14,17 +14,17 @@
     <!-- HEADER START -->
     <header class="container-fluid headerBar">
         <div class="row">
-            <a class="col-2 ml-1 mainLogoWhite logoWhite"  href="./"><img class=" logo-img" src="./img/logo.png" alt="logo"></a>
-            <a class="col-2 ml-1 mainLogoBlack logoBlack" href="./"><img class=" logo-img" src="./img/logo-blakc.png" alt="logo"></a>
+            <a class="col-2 ml-1 logoWhite"  href="./"><img class=" logo-img" src="./img/logo.png" alt="logo"></a>
+            <a class="col-2 ml-1 logoBlack" href="./"><img class=" logo-img" src="./img/logo-blakc.png" alt="logo"></a>
             <nav class="hidden visible-lg meniu-nav">
-                <a class="items item1 meniu-bar-selector" href="#top">Home</a>
-                <a class="items item2 meniu-bar-selector" href="#about">About</a>
-                <a class="items item3 meniu-bar-selector" href="#">Services</a>
-                <a class="items item4 meniu-bar-selector" href="#education">Education</a>
-                <a class="items item5 meniu-bar-selector" href="#">Work</a>
-                <a class="items item6 meniu-bar-selector" href="#">Client</a>
-                <a class="items item7 meniu-bar-selector" href="#">Blog</a>
-                <a class="items item8 meniu-bar-selector" href="#">Contact</a>
+                <a class="meniu-bar-selector" href="#top">Home</a>
+                <a class="meniu-bar-selector" href="#about">About</a>
+                <a class="meniu-bar-selector" href="#">Services</a>
+                <a class="meniu-bar-selector" href="#education">Education</a>
+                <a class="meniu-bar-selector" href="#">Work</a>
+                <a class="meniu-bar-selector" href="#">Client</a>
+                <a class="meniu-bar-selector" href="#">Blog</a>
+                <a class="meniu-bar-selector" href="#">Contact</a>
             </nav>
             <a class="hidden-lg meniu-bar-icon meniu-nav" href="#"><i class="fa fa-bars" aria-hidden="true"></i></a>
             </div>
@@ -296,7 +296,7 @@
         <div class="container hire-me-details">
             <div class="row hire-me">
                 <h2>I am available for freelancer.</h2>
-                <button class="btn btn-hover">Hire me!</button>            
+                <button class="btn">Hire me!</button>            
             </div>
         </div>
     </div>
@@ -406,7 +406,7 @@
     <!-- FOOTER END -->
 
     <!-- BACK TO TOP BUTTON START-->           
-    <a href="#" class="btn btn-black backToTopButton-hide"><i class="fa fa-angle-up" aria-hidden="true"></i></a>      
+    <a id="backToTop" href="#" class="btn btn-black backToTopButton-hide"><i class="fa fa-angle-up" aria-hidden="true"></i></a>      
     <!-- BACK TO TOP BUTTON END-->
 
     <script src="./js/mainHome.js" type="module" defer></script>
